Review Changes

+++ Data/models/DataController.swift
@@ -31,13 +31,16 @@ public class DataController: NSObject {
     }
     
     public func migrateToNewPathIfNeeded() {
-        if FileManager.default.fileExists(atPath: oldStoreURL.path) {
+        func sqliteFiles(from url: URL, dbName: String) throws -> [URL] {
+            return try FileManager.default.contentsOfDirectory(at: url, includingPropertiesForKeys: nil, options: []).filter({$0.lastPathComponent.hasPrefix(dbName)})
+        }
+        if FileManager.default.fileExists(atPath: oldStoreURL.path) && !storeExists() {
             do {
                 try migrationContainer.persistentStoreCoordinator.addPersistentStore(ofType: NSSQLiteStoreType, configurationName: nil, at: oldStoreURL, options: nil)
                 if let oldStore = migrationContainer.persistentStoreCoordinator.persistentStore(for: oldStoreURL) {
                     try migrationContainer.persistentStoreCoordinator.migratePersistentStore(oldStore, to: storeURL, options: nil, withType: NSSQLiteStoreType)
                     try migrationContainer.persistentStoreCoordinator.destroyPersistentStore(at: oldStoreURL, ofType: NSSQLiteStoreType, options: nil)
-                    try oldStoreURL.deletingLastPathComponent().sqliteFiles(dbName: DataController.databaseName).forEach(FileManager.default.removeItem)
+                    try sqliteFiles(from: oldStoreURL.deletingLastPathComponent(), dbName: DataController.databaseName).forEach(FileManager.default.removeItem)
                 }
             } catch {
                 log.error(error)
@@ -163,6 +166,7 @@ public class DataController: NSObject {
         return queue
     }()
     
+    /// Warning! Please use storeURL. oldStoreURL is for migration purpose only.
     private let oldStoreURL: URL = {
         let urls = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)
         guard let docURL = urls.last else {
@@ -196,12 +200,5 @@ public class DataController: NSObject {
         let backgroundContext = DataController.sharedInMemory.container.newBackgroundContext()
         backgroundContext.mergePolicy = NSMergePolicy.mergeByPropertyStoreTrump
         return backgroundContext
-    }
-}
-
-fileprivate extension URL {
-    
-    func sqliteFiles(dbName: String) throws -> [URL] {
-        return try FileManager.default.contentsOfDirectory(at: self, includingPropertiesForKeys: nil, options: []).filter({$0.lastPathComponent.hasPrefix(dbName)})
     }
 }
